+++ src/configs/rollup.es6.config.js
@@ -22,6 +22,7 @@ const process = require('process')
 const babel = require('@rollup/plugin-babel').babel
 const babelPluginClassProperties = require('@babel/plugin-proposal-class-properties')
 const babelPresetTypescript = require('@babel/preset-typescript')
+const babelPresetEnv = require('@babel/preset-env')
 const resolve = require('@rollup/plugin-node-resolve').nodeResolve
 const commonjs = require('@rollup/plugin-commonjs')
 const alias = require('@rollup/plugin-alias')
@@ -62,10 +63,20 @@ module.exports = {
         '~': path.resolve(process.cwd(), 'node_modules/'),
       },
     }),
-    resolve({ extensions, mainFields: ['module', 'main', 'browser'] }),
+    resolve({ extensions, mainFields: buildHelpers.getResolveConfigForBundlers() }),
     commonjs({ sourceMap: false }),
     babel({
-      presets: [[babelPresetTypescript]],
+      presets: [
+        [
+          babelPresetEnv,
+          {
+            targets: {
+              safari: '12.0',
+            },
+          },
+        ],
+        [babelPresetTypescript],
+      ],
       extensions,
       babelHelpers: 'bundled',
       plugins: [babelPluginClassProperties],